Dodano sprawdzanie zalogowania użytkownika przed wykonaniem operacji w komponentach StatusIcon, StatusBtn, StarRating, DeleteStatusOption.

+++ src/app/delete-status-option/delete-status-option.component.ts
@@ -1,5 +1,7 @@
 import {Component, Input} from '@angular/core';
 import {BookStatesService} from "../../services/BookStatesService";
+import {AuthService} from "../../services/AuthService";
+import {Router} from "@angular/router";
 
 @Component({
   selector: 'app-delete-status-option',
@@ -12,11 +14,18 @@ export class DeleteStatusOptionComponent {
   @Input() book_id : number = 0;
   @Input() status : string = ''
 
-  constructor(private bookStatesService : BookStatesService) {
+  constructor(private bookStatesService : BookStatesService, private authService : AuthService, private router : Router) {
   }
 
   deleteFromStatus() {
-    this.bookStatesService.DeleteBookState(this.book_id).subscribe(() => window.location.reload())
+    this.authService.GetUserRole().subscribe(role => {
+      if (role === null) {
+        this.router.navigate(['/login']);
+      }
+      else {
+        this.bookStatesService.DeleteBookState(this.book_id).subscribe(() => window.location.reload())
+      }
+    })
   }
 
 }
